last edits and norm
