+++ .vscode/settings.json
@@ -1,3 +1,4 @@
 {
-    "editor.defaultFormatter": null
+    "editor.defaultFormatter": "keyhr.42-c-format",
+    "editor.formatOnSave": true
 }
