+++ Makefile
@@ -6,22 +6,24 @@
 #    By: yalechin <yalechin@student.42.fr>          +#+  +:+       +#+         #
 #                                                 +#+#+#+#+#+   +#+            #
 #    Created: 2024/02/24 17:27:32 by yalechin          #+#    #+#              #
-#    Updated: 2024/10/26 13:34:21 by yalechin         ###   ########.fr        #
+#    Updated: 2024/11/02 13:24:54 by yalechin         ###   ########.fr        #
 #                                                                              #
 # **************************************************************************** #
 
 NAME = cub3d 
-CFILES = cub3d.c \
-			minimap.c \
-			map_parsing.c \
-			error.c \
-			map_setting_utils.c \
-			movement.c \
-			ray_casting.c \
-			exit.c \
-			ft_utils.c \
-			render_wall.c \
-			map_check.c
+CFILES = srcs/cub3d.c \
+			srcs/minimap.c \
+			srcs/map_parsing.c \
+			srcs/error.c \
+			srcs/map_setting_utils.c \
+			srcs/movement.c \
+			srcs/ray_casting.c \
+			srcs/exit.c \
+			srcs/ft_utils.c \
+			srcs/render_wall.c \
+			srcs/map_check.c \
+			srcs/map_check2.c \
+			srcs/render.c
 OFILES = $(CFILES:.c=.o)
 
 FLAGS = -O3 -Wall -Wextra -Werror -g
