+++ maps/map2.cub
