+++ maps/small2.cub
