+++ maps/small_map.cub
