+++ srcs/ft_utils.c
@@ -6,28 +6,28 @@
 /*   By: yalechin <yalechin@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/19 16:23:34 by yalechin          #+#    #+#             */
-/*   Updated: 2024/10/19 17:43:10 by yalechin         ###   ########.fr       */
+/*   Updated: 2024/11/02 12:53:47 by yalechin         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "cub3d.h"
+#include "../incl/cub3d.h"
 
-//better pixel put 
-void	ft_pixel_put(t_mlx *mlx, int x, int y, int color)	
+// better pixel put
+void	ft_pixel_put(t_mlx *mlx, int x, int y, int color)
 {
-	if (x < 0) 
+	if (x < 0)
 		return ;
 	else if (x >= S_W)
 		return ;
-	if (y < 0) 
+	if (y < 0)
 		return ;
 	else if (y >= S_H)
 		return ;
-	mlx_put_pixel(mlx->img, x, y, color); 
+	mlx_put_pixel(mlx->img, x, y, color);
 }
 
 // normalize the angle
-float	angle_nor(float angle)	
+float	angle_nor(float angle)
 {
 	if (angle < 0)
 		angle += (2 * M_PI);
